Add google authentication

+++ controllers/api/v1/post_api.js
@@ -22,20 +22,19 @@ module.exports.destroy = async function (req, res) {
   try {
     const post = await Post.findById(req.params.id);
 
-    // if (post.user.toString() === req.user.id.toString()) {
+    if (post.user.toString() === req.user.id.toString()) {
+      post.deleteOne();
 
-    post.deleteOne();
+      await Comment.deleteMany({ post: req.params.id });
 
-    await Comment.deleteMany({ post: req.params.id });
-
-    return res.status(200).json({
-      message: "Post and associated comments deleted successfully",
-    });
-
-    // } else {
-    //   req.flash("error", "You can't Delete this Post!");
-    //   return res.redirect("back");
-    // }
+      return res.status(200).json({
+        message: "Post and associated comments deleted successfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "You can not delete the posts",
+      });
+    }
   } catch (err) {
     console.log("*******", err);
     return res.json(500, {
